Added an Add button the SnapsViewController.
Added a new Picture View Controller.

Added a swift file the new PictureViewController.

Added a segue to switch from the SnapsViewController to the PictureViewControlle.

Added a Camera button.

Added an ImageView.

Added a Text Field for the Description.

Added a button and set its tile to Next.

Added three @IBOutlet for (imageView, descriptionTextField and nextButton)

Added two @IBAction for (cameraTapped and nextTapped).