Changed the image format to JPG with ratio of 0.1 and assign it to the imageData constant.
Set the name of the image to NSUUID and uploaded it to Firebase.

Set downloadURL constant to metadata?.downloadURL().

Printed the value the downloadURL constant.

Added a swift file for the SelectUserViewController.

Added the UITableViewDataSource and UITableViewDelegate to class of class SelecUserViewController.

Added a tableView.

Added an @IBOtlet for the tableView.

Set the dataSource and delegate of the tableView to self.

Added the numberOfRowsInSection and cellForRowAt indexPath function.

Returned 0 for the numberOfRowsInSection function.

Returned UITableViewCell () for the cellForRowAt indexPath function.

Added the following line of code to add data to the Database of Firebase.