Added a Class named User which has two attributes (email and uid)
Added a property variable named users of an array of User class type and set its value to [].

Pulled the data with the child “users” from the Database of Firebase.

Printed content of snapshot.

Declared a local constant named user of type User class.

Stored the UUID value of snapshot into the user.uid  attribute.

Stored the the email value of snapshot into the user.email attribute.

Added the content of the local user constant to the users property variable.

Reloaded the data of the tableView.

Added the numberOfRowsInSection function and returned the count value of users variable.

Added the cellForRowAt indexPath function and returned value of cell.