Added the print command to debug the Picture View Controller code.
Updated the didSelectRowAt indexPath function for the Select User View Controller to move back to the root view controller after choosing an email account.

Added the didSelectRowAt indexPath function for the Snaps View Controller to move back to 1) set the local snap constant to value of snaps property variable array at current indexPath.row 2) perform segue to View Snap View Controller.

Updated the didSelectRowAt indexPath function for the Snaps View Controller to move back to 1) set the local snap constant to value of snaps property variable array at current indexPath.row 2) perform segue to View Snap View Controller.

Updated the prepare(for segue function for the Snaps View Controller to move back to 1) set the local nextVC constant equal to the segue destination of the View Snap View Controller 2) set the value of snap property variable to value of the sender as Snap.

Added a Swift file the ViewSnapViewController.

Added two @IBOutlet for the snapImageView and snapDescriptionLable.

Added a property variable name snap of type Snap class.

Set the text value of the snapDescriptionLable to the value of snap.snapImageDescription.