Added a new attribute named key for the Snap class.
Set the key attribute value of snap to equal the value of key for snapshot.

Added line of code from Firebase API for remove the corsponding child of the tapped snap.

The rest of the code is similar to the child adding code.

Imported the Firebase API in the View Snap View Controller.

Added the viewWillDisappear function.

Printed “Goodbye” for testing.

Added line of code from Firebase API for remove the current viewed snap when the user tap the back button..