Changed how the print command should do the output.
Passed variable value to the signIn with email function for Firebase instead of using  emailTextField.text! and passwordTextField.text! directly.

Added a function named removeSnap within the Swift file of Snaps class to encapsulate the code required to remove the the viewed snap before moving back to the Snaps View Controller.

Updated the viewWillDisappear function with removeSnap for clean coding.